flask/requirements.txt
Flask==2.2.2
pydantic==1.10.6
requests==2.25.1
pymongo==4.3.3
uvicorn==0.21.0
jinja2==3.1.2
python-dotenv==0.21.0
httpx==0.23.3
Flask-Pydantic==0.11.0